qrb2210-rb1: Update partition table to support edk2/UEFI
The longer term plan for RB1 is to add support for full edk2/UEFI stack
on RB1 similar to what we have on RB3G2. So let's update the partition
table in order to allow us to migrate from current ABL based boot flow
to rather UEFI based boot flow.

The GUID's for *_b partitions are updated as well as new UEFI A/B and
UEFI DTB A/B partitions are added. The recoveryinfo partition is also
added required for storage of trial boot counters. It brings
compatibility with RB3G2 atleast from partition table layout
perspective.

Signed-off-by: Sumit Garg <sumit.garg@oss.qualcomm.com>