Fix partition sector rounding behavior
Previously, when a partition size contained decimals, the calculation of memory
size did not correctly round up to the next whole sector. This could result in
insufficient space allocation. This change ensures that the partition memory
size does not contain decimals and the number of sectors is always rounded up
for accurate space usage.

Signed-off-by: xiaomzh <xiaomzh@qti.qualcomm.com>