Makefile: use actual conf file to define the targets
The makefile targets are defined mainly as a wildcard of
$(wildcard platforms/*) and $(platform)/partitions.xml. They both
define the location of the 'output' file (the XML file being the
output file), and they hardcode the platform folder. Together they
simply assume that there can be only one partitions.conf file inside
the platform folder which is ./platforms/<board>.

However for each board we want to be able to support multiple board
configurations, such as a board with either an eMMC or a UFS disk, or
perhaps a board that comes with a 16GB eMMC or 32 GB eMMC. So ideally
we want to support:

 ./platforms/<board>/emmc/paritions.conf
 ./platforms/<board>/emmc-32gb/paritions.conf
 ./platforms/<board>/ufs/paritions.conf

With this patch, we are looking for partitions.conf file, and we set
the makefile targets based on that. We essentially introduce support
for the following pattern:

 ./platforms/<board>/partitions.conf
 ./platforms/<board>/<variant>/partitions.conf

Similarly, make sure we are looking for contexts.xml templates in both
folders:

 ./platforms/<board>/contents.xml.in
 ./platforms/<board>/<variant>/contents.xml.in

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>