get_contents: use list instead of set for the partitions
Running get_contents multiple times with the same input files generate
different output files. This is because the list of partitions from
partitions.xml are parsed and stored in a set data structure, which is
unsorted. Looping on the set will return different output.

Convert the set into a list instead which leads to:
1. the output is sorted, and thus the generated file is stable
2. the order from partitions.xml is preserved

While making this change, I decided to put dict elements in the list
instead of a list, it's easier to refer to field by name than by
index.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>