gen_partition: skip physical_partition with no partition
gen_partition.py assumes there can be up to 5 LUNs, and generate an
XML file with 5 physical_partition entries. For QCS615, there are only
4 LUNs, and we end up generating partitions.xml with an extra, empty
physical_partition entry.

The right thing would probably be to parse partition.conf and detect
how many LUNs we have in there, but a simpler solution is to keep
trying 5 LUNs but only create the physical_partition entry once we
find at least one partition, which is what this patch does. It should
fix the problem we've noticed in #42.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@oss.qualcomm.com>